+++ x/metadata/keeper/keeper.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"fmt"
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -117,5 +118,22 @@ func (k Keeper) CreateAccountForKey(ctx sdk.Context, addr sdk.AccAddress, pubKey
 		}
 	}
 	k.authKeeper.SetAccount(ctx, account)
+	return nil
+}
+
+// ValidateAllOwnersAreSigners makes sure that all entries in the existingOwners list are contained in the signers list.
+func (k Keeper) ValidateAllOwnersAreSigners(existingOwners []string, signers []string) error {
+	for _, owner := range existingOwners {
+		found := false
+		for _, signer := range signers {
+			if owner == signer {
+				found = true
+				break
+			}
+		}
+		if !found {
+			return fmt.Errorf("missing signature from existing owner %s; required for update", owner)
+		}
+	}
 	return nil
 }
